requirements.txt
tensorflow==2.15.0
pillow
matplotlib
numpy
flask
flask-cors
